script: create function to verify if profile is bot
This is the time expensive way to verify bot. It simply calls the
PegaBot's API and verify if the probability to be a bot is over 70%.

It will be used to let user click and verify if a profile on Twitter is
a bot or not, not to filter the posts.

Signed-off-by: Lucas Cardoso <lucas.cardoso.santos@usp.br>